refactor(tocco-util): move path/regex handling to util
Refs: TOCDEV-5060
