+++ packages/admin/package.json
@@ -9,7 +9,6 @@
   },
   "dependencies": {
     "js-cookie": "^3.0.1",
-    "path-to-regexp": "^3.0.0",
     "react-burger-menu": "^2.6.11"
   },
   "devDependencies": {
