+++ packages/admin/src/routes/entities/utils/url.js
@@ -1,4 +1,4 @@
-import pathToRegexp from 'path-to-regexp'
+import {route} from 'tocco-util'
 
 const entityUrlRegex = '/e/:entity/:key?/:relation*/:view(list|detail|edit|create|relations|action)/:actionId?'
 const actionUrlRegex = '/e/action/:actionId'
@@ -9,27 +9,8 @@ const actionUrlRegex = '/e/action/:actionId'
  * @param pathname {String} url / path starting with /e/
  */
 export const getPathInfo = pathname => {
-  const pathParts = []
-  const re = pathToRegexp(entityUrlRegex, pathParts)
-  const res = re.exec(pathname)
+  const entityParams = route.extractParamsFromPath(entityUrlRegex, pathname)
+  const actionParams = route.extractParamsFromPath(actionUrlRegex, pathname)
 
-  if (res !== null) {
-    return pathParts.reduce((acc, pathPart, idx) => {
-      return {
-        ...acc,
-        [pathPart.name]: res[idx + 1]
-      }
-    }, {})
-  }
-
-  const actionPathRegex = pathToRegexp(actionUrlRegex, pathParts)
-  const parseResult = actionPathRegex.exec(pathname)
-
-  if (parseResult !== null) {
-    return {
-      actionId: parseResult[1]
-    }
-  }
-
-  return null
+  return entityParams || actionParams
 }
