+++ packages/tocco-util/src/route/index.js
@@ -1,3 +1,4 @@
+import {extractParamsFromPath} from './path'
 import {loadRoute} from './route'
 import Router from './Router'
 import RouteWithSubRoutes from './RouteWithSubRoutes'
@@ -5,5 +6,6 @@ import RouteWithSubRoutes from './RouteWithSubRoutes'
 export default {
   loadRoute,
   Router,
-  RouteWithSubRoutes
+  RouteWithSubRoutes,
+  extractParamsFromPath
 }
